move AddReissuanceArgs and AddIssuanceArgs
